added `make dogfood` The new makefile target `dogfood` supports a workflow for developers to use Hjkl for writing Hjkl code. Running `make dogfood` will build and run Hjkl with all source files included on the disk image. After closing the virtual machine, these files will be updated from the disk image, reflecting any changes made within the virtual environment. Backups of all source files are made before running.
The dogfooding process is currently highly flawed! For one, if an edit to the source code causes a compilation failure (highly common during development, obviously), then the code has to be fixed "manually" on the physical file as normal. The backup process also leaves something to be desired; would it be crazy to have the backup folder be a git repository?

Hjkl is not yet mature enough to be dogfooded, but that observation is perhaps the whole point of dogfooding, and it was fun to implement nonetheless.